refactor: move ProjectQueryFilterBuilder to package
